pyproject.toml
[tool.poetry]
name = "auditory-deepdream"
version = "0.1.0"
description = ""
authors = ["Andrea Spiteri <aspit002@gold.ac.uk>"]
packages = [
    { include = "ai_tools", from = "library" },
    { include = "deep_dream", from = "library" },
    { include = "utils", from = "library" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
tensorflow = "^2.7.0"
numpy = "^1.20.0"
librosa = "^0.8.1"
matplotlib = "^3.5.1"
tqdm = "^4.62.3"
kapre = "^0.3.7"
jupyter = "^1.0.0"
pandas = "^1.4.1"
PyYAML = "^6.0"
aim = "^3.6.0"
bayesian-optimization = "^1.2.0"
opencv-python = "^4.5.5"
scikit-maad = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
